Bootstrap and themobile first design
At the beginning of Bootstrap, there was no concept of mobile first, so it was made to work for designing responsive web pages. However, with the version 3 of the framework, the concept of mobile first was very solid in the community.

For doing this, the whole code of the scaffolding system was rewritten to become mobile first from the start. They decided to reformulate how to set up the grid instead of just adding mobile styles. This made a great impact on compatibility between versions older than 3, but was crucial for making the framework even more popular.

To ensure the proper rendering of the page, set the correct viewport at the <head> tag:

<meta name="viewport" content="width=device-width,
  initial-scale=1">